[fix] getToday()をローカル時刻基準に統一し共通化
record.tsx/recommend.txtでtoISOString()（UTC基準）を使っていたため、
日本時間0:00〜8:59に記録すると日付が1日ずれて履歴に表示されなかった。
全画面のgetToday()をローカル時刻基準のlib/date.tsに統一。

Closes #79
