+++ apps/mobile/lib/date.ts
@@ -1,0 +1,8 @@
+/** Returns today's date in YYYY-MM-DD format based on local timezone. */
+export function getToday(): string {
+  const d = new Date();
+  const year = d.getFullYear();
+  const month = String(d.getMonth() + 1).padStart(2, "0");
+  const day = String(d.getDate()).padStart(2, "0");
+  return `${year}-${month}-${day}`;
+}
